// services/api/dbschema/interfaces.ts
export namespace std {
  export interface BaseObject {
    "id": string;
  }
  export interface $Object extends BaseObject {}
  export interface FreeObject extends BaseObject {}
  export type JsonEmpty = "ReturnEmpty" | "ReturnTarget" | "Error" | "UseNull" | "DeleteKey";
}
// ...
export interface Feed extends std.$Object {
  "feedName": string;
  "feedType": string;
  "feedId": string;
  "favourited"?: boolean | null;
  "feedNameLC": string;
  "lastUpdated"?: number | null;
  "updateCheck_LastPostSeen"?: string | null;
  "posts": Post[];
  "tags": Tag[];
}
export interface Log extends std.$Object {
  "level": string;
  "createdAt": number;
  "error"?: string | null;
  "message"?: string | null;
  "other"?: unknown | null;
  "otherAsStr"?: string | null;
  "service"?: string | null;
}
export interface Post extends std.$Object {
  "feedType": string;
  "postId": string;
  "uniqueId": string;
  "feedName": string;
  "timestamp": number;
  "couldNotDownload"?: boolean | null;
  "downloadError"?: string | null;
  "downloadedMedia"?: string[] | null;
  "downloadedMediaCount"?: number | null;
  "mediaDownloadTries"?: number | null;
  "mediaHasBeenDownloaded"?: boolean | null;
  "mediaUrl": string;
  "postMediaImagesHaveBeenProcessed"?: boolean | null;
  "postMediaImagesProcessingError"?: string | null;
  "postThumbnailsCreated"?: boolean | null;
  "postUrl": string;
  "score": number;
  "title": string;
  "feed": Feed;
  "tags": Tag[];
}
export interface Settings extends std.$Object {
  "archiveImageCompressionQuality": number;
  "imageCompressionQuality": number;
  "maxImageWidthForNonArchiveImage": number;
  "numberImagesProcessAtOnce": number;
  "numberMediaDownloadsAtOnce": number;
  "uniqueId": string;
  "updateAllDay": boolean;
  "updateEndingHour": number;
  "updateStartingHour": number;
}
export interface Tag extends std.$Object {
  "favourited"?: boolean | null;
  "tag": string;
  "feeds": Feed[];
  "posts": Post[];
}